backend/requirements.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
gunicorn
pytest
pytest-django
flake8
black
BioPython
tensorflow
torch
torchvision